Optimize Boundary Check in BinarySearch Method
This commit enhances the performance of the BinarySearch method by eliminating unnecessary boundary checks. The condition if ((uint)candidate.Length > (uint)prefix.Length) ensures that the index access candidate[prefix.Length] is safe, and the JIT compiler can now safely elide the boundary check, improving execution efficiency.

+++ src/Mvc/Mvc.Core/src/ModelBinding/PrefixContainer.cs
@@ -212,7 +212,7 @@ public class PrefixContainer
                 // which should not qualify as a prefix match. Therefore, we will continue searching
                 // for valid matches in this case.
 
-                if (candidate.Length > prefix.Length)
+                if ((uint)candidate.Length > (uint)prefix.Length)
                 {
                     var c = candidate[prefix.Length];
                     if (c == '.' || c == '[')
